requirements.txt
opencv-python>=4.5.0
numpy>=1.20.0
python-dotenv>=0.19.0
